scaffold is changed to React 16
