+++ index.html
@@ -19,8 +19,9 @@
   <div id="TingleDemo"></div>
   <script src="http://g.alicdn.com/platform/nw_hybrid_jssdk/0.1.30/loader.js"></script>
   <!-- <script src="./node_modules/fastclick/lib/fastclick.js"></script> -->
-  <script src="./node_modules/react/dist/react-with-addons.js"></script>
-  <script src="./node_modules/react-dom/dist/react-dom.js"></script>
+  <script src="./node_modules/babel-polyfill/dist/polyfill.min.js"></script>
+  <script src="./node_modules/react/umd/react.development.js"></script>
+  <script src="./node_modules/react-dom/umd/react-dom.development.js"></script>
   <script src="./dist/demo.js"></script>
 </body>
