+++ src/Slot/tpls/Header.jsx
@@ -1,7 +1,7 @@
-const React = require('react');
-const classnames = require('classnames');
-const Context = require('../../Context');
-const PropTypes = require('prop-types');
+import React from 'react';
+import PropTypes from 'prop-types';
+import classnames from 'classnames';
+import Context from '../../Context';
 
 class SlotHeader extends React.Component {
   static displayName = 'SlotHeader'
